Add UI with state tracking and iteration start logic for LLM-assistant messages
Implemented UI that tracks the state of messages between the LLM and the assistant.

Added logic in the UI to send requests to start message iteration.

Also minor refactoring of filenames
